+++ src/Resource/Envelope.php
@@ -31,7 +31,7 @@ class Envelope extends BaseResource
     /** @deprecated Use $expirationPeriod instead */
     public int $expiration;
 
-    public TimePeriod $expirationPeriod;
+    public Period $expirationPeriod;
 
     public string $expirationTimeOfDay;